Add "Reload with Encoding" menu items and command
When ConvertToUTF8 failed to detect the encoding correctly, either
because of insufficient bytes from file or improper settings, we need
to manually specify the source file encoding. However, ConvertToUTF8
doesn't have this kind of settings/utility.

A simple test case is a GBK encoded text file with only two Chinese
characters "回家" (it may contain other ASCII characters). EUC-JP will
be detected as the source encoding and the Chinese characters are rendered
as "指社". As far as I have tested, "File" -> "Set File Encoding to"
won't help in this case.

Now, you have two choices to manually specify the source encoding when
it goes wrong:
1) "File" -> "Reload with Encoding" -> Choose the correct file encoding
2) Ctrl+Shift+P to activate the "Command Palette..." and search for
   "ConvertToUTF8: Reload with Encoding", select it and then choose
   the correct file encoding.